pyproject.toml
[project]
name = "the-agent-sandbox-taxonomy"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
]
